requirements.txt
annotated-types==0.7.0
anyio==4.5.2
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
jiter==0.9.1
openai==1.84.0
outcome==1.3.0.post0
packaging==25.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PySocks==1.7.1
python-dotenv==1.0.1
python-multipart==0.0.20
python-telegram-bot==21.6
PyYAML==6.0.2
requests==2.32.3
selenium==4.27.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
starlette==0.44.0
tqdm==4.67.1
trio==0.27.0
trio-websocket==0.12.2
typing_extensions==4.13.2
urllib3==2.2.3
uvicorn==0.33.0
watchfiles==0.24.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==13.1
wsproto==1.2.0
sqlalchemy
dateparser
jinja2
alembic
xendit-python
requests
psycopg2-binary